Fix chart data rendering
